Add script to back up repositories in batch
A script to clone and create a compressed zip file of all repos we maintain. This is consistent with the best practices for the GA Software Engineering Standard.

For the given list of repositories, the latest tag is cloned and saved to a compressed zip file, including the date and tag name for the repository. This doesn't clone the branches of the repository (though that could be done my adding the url of the branch).

The `backup_repos.cmd` can be used in a Windows Task Schedule task to set the script to run on regular intervals (e.g. weekly)